requirements.txt
blacksheep==1.2.7
asyncpg==0.26.0
SQLAlchemy==1.4.41
pytest==7.1.3
uvicorn==0.18.3
alembic==1.8.1
pyjwt==2.5.0